chore(plugins): Swift plugins
