+++ crates/facet_generate/src/generation/json/swift.rs
@@ -1,0 +1,739 @@
+//! `EmitterPlugin<Swift>` implementation for the [`JsonPlugin`].
+//!
+//! Provides JSON-specific imports, feature helper snippets, and full
+//! type-body generation (serialize / deserialize methods + wrappers) for
+//! Swift code generation.
+//!
+//! # What this plugin handles
+//!
+//! | Extension point | What it provides |
+//! |---|---|
+//! | `imports` | `import Serde` |
+//! | `module_helpers` | Feature helper snippets (`ListOfT`, `SetOfT`, …) |
+//! | `has_type_body` | Always `true` |
+//! | `type_body` | `serialize` / `deserialize` methods + `jsonSerialize` / `jsonDeserialize` wrappers |
+//!
+//! # JSON struct-field tuple serialisation
+//!
+//! For a struct field `foo: (A, B)`, JSON wraps the elements in an extra
+//! container-depth push/pop, using the field name (without `self.`) as the
+//! tuple receiver:
+//! ```swift
+//! try serializer.increase_container_depth()   // tuple-level
+//! try serializer.serialize_a(value: foo.0)
+//! try serializer.serialize_b(value: foo.1)
+//! try serializer.decrease_container_depth()
+//! ```
+//! Combined with the struct-level push/pop this produces double nesting.
+
+use std::collections::BTreeMap;
+use std::io;
+
+use heck::ToLowerCamelCase as _;
+use indoc::writedoc;
+
+use crate::generation::{
+    CodeGeneratorConfig, Feature,
+    indent::{IndentWrite, Newlines, with_block},
+    plugin::{EmitContext, EmitterPlugin},
+    swift::Swift,
+};
+use crate::reflection::format::{ContainerFormat, Format, Named, VariantFormat};
+
+use super::JsonPlugin;
+
+// ---------------------------------------------------------------------------
+// Inlined feature helper snippets (identical to the Bincode plugin)
+// ---------------------------------------------------------------------------
+
+const FEATURE_LIST_OF_T: &str = r"func serializeArray<T, S: Serializer>(
+    value: [T],
+    serializer: S,
+    serializeElement: (T, S) throws -> Void
+) throws {
+    try serializer.serialize_len(value: value.count)
+    for item in value {
+        try serializeElement(item, serializer)
+    }
+}
+
+func deserializeArray<T, D: Deserializer>(
+    deserializer: D,
+    deserializeElement: (D) throws -> T
+) throws -> [T] {
+    let length = try deserializer.deserialize_len()
+    var obj: [T] = []
+    for _ in 0..<length {
+        obj.append(try deserializeElement(deserializer))
+    }
+    return obj
+}
+";
+
+const FEATURE_MAP_OF_T: &str = r"func serializeMap<K, V, S: Serializer>(
+    value: [K: V],
+    serializer: S,
+    serializeEntry: (K, V, S) throws -> Void
+) throws {
+    try serializer.serialize_len(value: value.count)
+    for (key, value) in value {
+        try serializeEntry(key, value, serializer)
+    }
+}
+
+func deserializeMap<K: Hashable, V, D: Deserializer>(
+    deserializer: D,
+    deserializeEntry: (D) throws -> (K, V)
+) throws -> [K: V] {
+    let length = try deserializer.deserialize_len()
+    var obj: [K: V] = [:]
+    for _ in 0..<length {
+        let (key, value) = try deserializeEntry(deserializer)
+        obj[key] = value
+    }
+    return obj
+}
+";
+
+const FEATURE_OPTION_OF_T: &str = r"func serializeOption<T, S: Serializer>(
+    value: T?,
+    serializer: S,
+    serializeElement: (T, S) throws -> Void
+) throws {
+    if let value = value {
+        try serializer.serialize_option_tag(value: true)
+        try serializeElement(value, serializer)
+    } else {
+        try serializer.serialize_option_tag(value: false)
+    }
+}
+
+func deserializeOption<T, D: Deserializer>(
+    deserializer: D,
+    deserializeElement: (D) throws -> T
+) throws -> T? {
+    let tag = try deserializer.deserialize_option_tag()
+    if tag {
+        return try deserializeElement(deserializer)
+    } else {
+        return nil
+    }
+}
+";
+
+const FEATURE_SET_OF_T: &str = r"func serializeSet<T: Hashable, S: Serializer>(
+    value: Set<T>,
+    serializer: S,
+    serializeElement: (T, S) throws -> Void
+) throws {
+    try serializer.serialize_len(value: value.count)
+    for item in value {
+        try serializeElement(item, serializer)
+    }
+}
+
+func deserializeSet<T: Hashable, D: Deserializer>(
+    deserializer: D,
+    deserializeElement: (D) throws -> T
+) throws -> Set<T> {
+    let length = try deserializer.deserialize_len()
+    var obj: Set<T> = []
+    for _ in 0..<length {
+        obj.insert(try deserializeElement(deserializer))
+    }
+    return obj
+}
+";
+
+const FEATURE_TUPLE_ARRAY: &str = r"func serializeTupleArray<T, S: Serializer>(
+    value: [T],
+    serializer: S,
+    serializeElement: (T, S) throws -> Void
+) throws {
+    for item in value {
+        try serializeElement(item, serializer)
+    }
+}
+
+func deserializeTupleArray<T, D: Deserializer>(
+    deserializer: D,
+    size: Int,
+    deserializeElement: (D) throws -> T
+) throws -> [T] {
+    var obj: [T] = []
+    for _ in 0..<size {
+        obj.append(try deserializeElement(deserializer))
+    }
+    return obj
+}
+";
+
+// ---------------------------------------------------------------------------
+// EmitterPlugin implementation
+// ---------------------------------------------------------------------------
+
+impl EmitterPlugin<Swift> for JsonPlugin {
+    fn imports(&self, _config: &CodeGeneratorConfig) -> Vec<String> {
+        vec!["Serde".to_string()]
+    }
+
+    fn module_helpers(
+        &self,
+        w: &mut dyn IndentWrite,
+        config: &CodeGeneratorConfig,
+    ) -> io::Result<()> {
+        for feature in &config.features {
+            match feature {
+                Feature::OptionOfT => {
+                    writeln!(w)?;
+                    write!(w, "{FEATURE_OPTION_OF_T}")?;
+                }
+                Feature::ListOfT => {
+                    writeln!(w)?;
+                    write!(w, "{FEATURE_LIST_OF_T}")?;
+                }
+                Feature::SetOfT => {
+                    writeln!(w)?;
+                    write!(w, "{FEATURE_SET_OF_T}")?;
+                }
+                Feature::MapOfT => {
+                    writeln!(w)?;
+                    write!(w, "{FEATURE_MAP_OF_T}")?;
+                }
+                Feature::TupleArray => {
+                    writeln!(w)?;
+                    write!(w, "{FEATURE_TUPLE_ARRAY}")?;
+                }
+                _ => {}
+            }
+        }
+        Ok(())
+    }
+
+    fn has_type_body(&self, _ctx: &EmitContext) -> bool {
+        true
+    }
+
+    fn type_body(&self, w: &mut dyn IndentWrite, ctx: &EmitContext) -> io::Result<()> {
+        let name = ctx.name();
+        if let ContainerFormat::Enum(variants, _) = ctx.container.format {
+            write_enum_type_body(w, name, variants)
+        } else {
+            write_struct_type_body(w, name, &ctx.fields())
+        }
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Struct type body
+// ---------------------------------------------------------------------------
+
+fn write_struct_type_body(
+    w: &mut dyn IndentWrite,
+    name: &str,
+    fields: &[Named<Format>],
+) -> io::Result<()> {
+    writeln!(w)?;
+    write!(
+        w,
+        "public func serialize<S: Serializer>(serializer: S) throws "
+    )?;
+    with_block(w, Newlines::BOTH, |w| {
+        push_serializer(w)?;
+        for field in fields {
+            let fname = field.name.to_lower_camel_case();
+            write_field_serialize(w, &fname, &field.value)?;
+        }
+        pop_serializer(w)
+    })?;
+    write_json_serialize(w)?;
+
+    writeln!(w)?;
+    write!(
+        w,
+        "public static func deserialize<D: Deserializer>(deserializer: D) throws -> {name} "
+    )?;
+    with_block(w, Newlines::BOTH, |w| {
+        push_deserializer(w)?;
+        for field in fields {
+            let fname = field.name.to_lower_camel_case();
+            write_field_deserialize(w, &fname, &field.value)?;
+        }
+        pop_deserializer(w)?;
+        write!(w, "return {name}(")?;
+        for (i, field) in fields.iter().enumerate() {
+            if i > 0 {
+                write!(w, ", ")?;
+            }
+            let fname = field.name.to_lower_camel_case();
+            write!(w, "{fname}: {fname}")?;
+        }
+        writeln!(w, ")")
+    })?;
+    write_json_deserialize(w, name)?;
+
+    Ok(())
+}
+
+/// Serialise one struct field, applying JSON tuple wrapping when the field
+/// type is a native tuple.
+fn write_field_serialize(w: &mut dyn IndentWrite, fname: &str, format: &Format) -> io::Result<()> {
+    match format {
+        // JSON wraps native tuple fields in an extra container-depth level.
+        // The receiver is the field name (no `self.` prefix) so elements are
+        // addressed as `fname.0`, `fname.1`, …
+        Format::Tuple(formats) => {
+            push_serializer(w)?;
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_serialize(w, fmt, &format!("{fname}.{i}"))?;
+            }
+            pop_serializer(w)
+        }
+        _ => write_format_serialize(w, format, &format!("self.{fname}")),
+    }
+}
+
+/// Deserialise one struct field, applying JSON tuple wrapping when the field
+/// type is a native tuple.
+fn write_field_deserialize(
+    w: &mut dyn IndentWrite,
+    fname: &str,
+    format: &Format,
+) -> io::Result<()> {
+    match format {
+        // JSON wraps native tuple fields in an extra container-depth level.
+        // Intermediate variables are named `fname0`, `fname1`, … (no `Field`
+        // infix, matching the original emitter behaviour).
+        Format::Tuple(formats) => {
+            push_deserializer(w)?;
+            for (i, fmt) in formats.iter().enumerate() {
+                let elem = format!("{fname}{i}");
+                write_format_deserialize(w, fmt, &elem)?;
+            }
+            write!(w, "let {fname} = (")?;
+            for i in 0..formats.len() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                write!(w, "{fname}{i}")?;
+            }
+            writeln!(w, ")")?;
+            pop_deserializer(w)
+        }
+        _ => write_format_deserialize(w, format, fname),
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Enum type body (identical structure to BincodePlugin)
+// ---------------------------------------------------------------------------
+
+fn write_enum_type_body(
+    w: &mut dyn IndentWrite,
+    name: &str,
+    variants: &BTreeMap<u32, Named<VariantFormat>>,
+) -> io::Result<()> {
+    writeln!(w)?;
+    write!(
+        w,
+        "public func serialize<S: Serializer>(serializer: S) throws "
+    )?;
+    with_block(w, Newlines::BOTH, |w| {
+        push_serializer(w)?;
+        write!(w, "switch self ")?;
+        with_block(w, Newlines::BOTH, |w| {
+            w.unindent();
+            for (i, variant) in variants.values().enumerate() {
+                write_variant_serialize_case(w, variant, i)?;
+            }
+            w.indent();
+            Ok(())
+        })?;
+        pop_serializer(w)
+    })?;
+    write_json_serialize(w)?;
+
+    writeln!(w)?;
+    write!(
+        w,
+        "public static func deserialize<D: Deserializer>(deserializer: D) throws -> {name} "
+    )?;
+    with_block(w, Newlines::BOTH, |w| {
+        writeln!(
+            w,
+            "let index = try deserializer.deserialize_variant_index()"
+        )?;
+        push_deserializer(w)?;
+        write!(w, "switch index ")?;
+        with_block(w, Newlines::BOTH, |w| {
+            w.unindent();
+            for (i, variant) in variants.values().enumerate() {
+                write_variant_deserialize_case(w, variant, i)?;
+            }
+            writeln!(
+                w,
+                r#"default: throw DeserializationError.invalidInput(issue: "Unknown variant index for {name}: \(index)")"#
+            )?;
+            w.indent();
+            Ok(())
+        })
+    })?;
+    write_json_deserialize(w, name)?;
+
+    Ok(())
+}
+
+// ---------------------------------------------------------------------------
+// Variant helpers (identical to BincodePlugin)
+// ---------------------------------------------------------------------------
+
+fn write_variant_serialize_case(
+    w: &mut dyn IndentWrite,
+    variant: &Named<VariantFormat>,
+    index: usize,
+) -> io::Result<()> {
+    let name = variant.name.to_lower_camel_case();
+    match &variant.value {
+        VariantFormat::Variable(_) => unreachable!("placeholders should not get this far"),
+        VariantFormat::Unit => {
+            writeln!(w, "case .{name}:")?;
+            w.indent();
+            writeln!(w, "try serializer.serialize_variant_index(value: {index})")?;
+            w.unindent();
+        }
+        VariantFormat::NewType(fmt) => {
+            writeln!(w, "case .{name}(let x):")?;
+            w.indent();
+            writeln!(w, "try serializer.serialize_variant_index(value: {index})")?;
+            write_format_serialize(w, fmt, "x")?;
+            w.unindent();
+        }
+        VariantFormat::Tuple(formats) => {
+            write!(w, "case .{name}(")?;
+            for (i, _) in formats.iter().enumerate() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                write!(w, "let x{i}")?;
+            }
+            writeln!(w, "):")?;
+            w.indent();
+            writeln!(w, "try serializer.serialize_variant_index(value: {index})")?;
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_serialize(w, fmt, &format!("x{i}"))?;
+            }
+            w.unindent();
+        }
+        VariantFormat::Struct(nameds) => {
+            write!(w, "case .{name}(")?;
+            for (i, named) in nameds.iter().enumerate() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                let field_name = named.name.to_lower_camel_case();
+                write!(w, "let {field_name}")?;
+            }
+            writeln!(w, "):")?;
+            w.indent();
+            writeln!(w, "try serializer.serialize_variant_index(value: {index})")?;
+            for named in nameds {
+                let field_name = named.name.to_lower_camel_case();
+                write_format_serialize(w, &named.value, &field_name)?;
+            }
+            w.unindent();
+        }
+    }
+    Ok(())
+}
+
+fn write_variant_deserialize_case(
+    w: &mut dyn IndentWrite,
+    variant: &Named<VariantFormat>,
+    index: usize,
+) -> io::Result<()> {
+    let name = variant.name.to_lower_camel_case();
+    writeln!(w, "case {index}:")?;
+    w.indent();
+    match &variant.value {
+        VariantFormat::Variable(_) => unreachable!("placeholders should not get this far"),
+        VariantFormat::Unit => {
+            pop_deserializer(w)?;
+            writeln!(w, "return .{name}")?;
+        }
+        VariantFormat::NewType(fmt) => {
+            write_format_deserialize(w, fmt, "x")?;
+            pop_deserializer(w)?;
+            writeln!(w, "return .{name}(x)")?;
+        }
+        VariantFormat::Tuple(formats) => {
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_deserialize(w, fmt, &format!("x{i}"))?;
+            }
+            pop_deserializer(w)?;
+            write!(w, "return .{name}(")?;
+            for (i, _) in formats.iter().enumerate() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                write!(w, "x{i}")?;
+            }
+            writeln!(w, ")")?;
+        }
+        VariantFormat::Struct(nameds) => {
+            for named in nameds {
+                let field_name = named.name.to_lower_camel_case();
+                write_format_deserialize(w, &named.value, &field_name)?;
+            }
+            pop_deserializer(w)?;
+            write!(w, "return .{name}(")?;
+            for (i, named) in nameds.iter().enumerate() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                let field_name = named.name.to_lower_camel_case();
+                write!(w, "{field_name}: {field_name}")?;
+            }
+            writeln!(w, ")")?;
+        }
+    }
+    w.unindent();
+    Ok(())
+}
+
+// ---------------------------------------------------------------------------
+// Encoding wrappers
+// ---------------------------------------------------------------------------
+
+fn write_json_serialize(w: &mut dyn IndentWrite) -> io::Result<()> {
+    writeln!(w)?;
+    writedoc!(
+        w,
+        r"
+        public func jsonSerialize() throws -> [UInt8] {{
+            let serializer = JsonSerializer.init();
+            try self.serialize(serializer: serializer)
+            return serializer.get_bytes()
+        }}
+        "
+    )
+}
+
+fn write_json_deserialize(w: &mut dyn IndentWrite, name: &str) -> io::Result<()> {
+    writeln!(w)?;
+    writedoc!(
+        w,
+        r#"
+        public static func jsonDeserialize(input: [UInt8]) throws -> {name} {{
+            let deserializer = JsonDeserializer.init(input: input);
+            let obj = try deserialize(deserializer: deserializer)
+            if deserializer.get_buffer_offset() < input.count {{
+                throw DeserializationError.invalidInput(issue: "Some input bytes were not read")
+            }}
+            return obj
+        }}
+        "#
+    )
+}
+
+// ---------------------------------------------------------------------------
+// Format serialisation helpers (identical to BincodePlugin)
+// ---------------------------------------------------------------------------
+
+fn write_format_serialize(
+    w: &mut dyn IndentWrite,
+    format: &Format,
+    value_expr: &str,
+) -> io::Result<()> {
+    match format {
+        Format::Variable(_) => unreachable!("placeholders should not get this far"),
+        Format::TypeName(_) => writeln!(w, "try {value_expr}.serialize(serializer: serializer)"),
+        Format::Option(inner) => {
+            write!(
+                w,
+                "try serializeOption(value: {value_expr}, serializer: serializer) "
+            )?;
+            with_block(w, Newlines::CLOSE, |w| {
+                writeln!(w, " value, serializer in")?;
+                write_format_serialize(w, inner, "value")
+            })
+        }
+        Format::Seq(inner) => {
+            write!(
+                w,
+                "try serializeArray(value: {value_expr}, serializer: serializer) "
+            )?;
+            with_block(w, Newlines::CLOSE, |w| {
+                writeln!(w, " item, serializer in")?;
+                write_format_serialize(w, inner, "item")
+            })
+        }
+        Format::Set(inner) => {
+            write!(
+                w,
+                "try serializeSet(value: {value_expr}, serializer: serializer) "
+            )?;
+            with_block(w, Newlines::CLOSE, |w| {
+                writeln!(w, " item, serializer in")?;
+                write_format_serialize(w, inner, "item")
+            })
+        }
+        Format::Map { key, value } => {
+            write!(
+                w,
+                "try serializeMap(value: {value_expr}, serializer: serializer) "
+            )?;
+            with_block(w, Newlines::CLOSE, |w| {
+                writeln!(w, " key, value, serializer in")?;
+                write_format_serialize(w, key, "key")?;
+                write_format_serialize(w, value, "value")
+            })
+        }
+        Format::Tuple(formats) => {
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_serialize(w, fmt, &format!("{value_expr}.{i}"))?;
+            }
+            Ok(())
+        }
+        Format::TupleArray { content, .. } => {
+            write!(
+                w,
+                "try serializeTupleArray(value: {value_expr}, serializer: serializer) "
+            )?;
+            with_block(w, Newlines::CLOSE, |w| {
+                writeln!(w, " item, serializer in")?;
+                write_format_serialize(w, content, "item")
+            })
+        }
+        primitive => {
+            let t = format!("{primitive:?}").to_lowercase();
+            writeln!(w, "try serializer.serialize_{t}(value: {value_expr})")
+        }
+    }
+}
+
+fn write_format_deserialize(w: &mut dyn IndentWrite, format: &Format, var: &str) -> io::Result<()> {
+    match format {
+        Format::Tuple(formats) if formats.len() > 1 => {
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_deserialize(w, fmt, &format!("{var}Field{i}"))?;
+            }
+            write!(w, "let {var} = (")?;
+            for i in 0..formats.len() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                write!(w, "{var}Field{i}")?;
+            }
+            writeln!(w, ")")
+        }
+        _ => {
+            write!(w, "let {var} = ")?;
+            write_deserialize_expr(w, format)?;
+            writeln!(w)
+        }
+    }
+}
+
+fn write_deserialize_expr(w: &mut dyn IndentWrite, format: &Format) -> io::Result<()> {
+    match format {
+        Format::Variable(_) => unreachable!("placeholders should not get this far"),
+        Format::TypeName(qtn) => {
+            let type_name = qtn.format(|ns| heck::AsUpperCamelCase(ns).to_string(), ".");
+            write!(w, "try {type_name}.deserialize(deserializer: deserializer)")
+        }
+        Format::Option(inner) => {
+            writeln!(
+                w,
+                "try deserializeOption(deserializer: deserializer) {{ deserializer in"
+            )?;
+            w.indent();
+            write_deserialize_expr(w, inner)?;
+            writeln!(w)?;
+            w.unindent();
+            write!(w, "}}")
+        }
+        Format::Seq(inner) => {
+            writeln!(
+                w,
+                "try deserializeArray(deserializer: deserializer) {{ deserializer in"
+            )?;
+            w.indent();
+            write_deserialize_expr(w, inner)?;
+            writeln!(w)?;
+            w.unindent();
+            write!(w, "}}")
+        }
+        Format::Set(inner) => {
+            writeln!(
+                w,
+                "try deserializeSet(deserializer: deserializer) {{ deserializer in"
+            )?;
+            w.indent();
+            write_deserialize_expr(w, inner)?;
+            writeln!(w)?;
+            w.unindent();
+            write!(w, "}}")
+        }
+        Format::Map { key, value } => {
+            writeln!(
+                w,
+                "try deserializeMap(deserializer: deserializer) {{ deserializer in"
+            )?;
+            w.indent();
+            write_format_deserialize(w, key, "key")?;
+            write_format_deserialize(w, value, "value")?;
+            writeln!(w, "return (key, value)")?;
+            w.unindent();
+            write!(w, "}}")
+        }
+        Format::Tuple(formats) if formats.len() == 1 => write_deserialize_expr(w, &formats[0]),
+        Format::Tuple(formats) => {
+            for (i, fmt) in formats.iter().enumerate() {
+                write_format_deserialize(w, fmt, &format!("field{i}"))?;
+            }
+            write!(w, "return (")?;
+            for i in 0..formats.len() {
+                if i > 0 {
+                    write!(w, ", ")?;
+                }
+                write!(w, "field{i}")?;
+            }
+            write!(w, ")")
+        }
+        Format::TupleArray { content, size } => {
+            writeln!(
+                w,
+                "try deserializeTupleArray(deserializer: deserializer, size: {size}) {{ deserializer in"
+            )?;
+            w.indent();
+            write_deserialize_expr(w, content)?;
+            writeln!(w)?;
+            w.unindent();
+            write!(w, "}}")
+        }
+        primitive => {
+            let t = format!("{primitive:?}").to_lowercase();
+            write!(w, "try deserializer.deserialize_{t}()")
+        }
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Depth tracking
+// ---------------------------------------------------------------------------
+
+fn push_serializer(w: &mut dyn IndentWrite) -> io::Result<()> {
+    writeln!(w, "try serializer.increase_container_depth()")
+}
+
+fn pop_serializer(w: &mut dyn IndentWrite) -> io::Result<()> {
+    writeln!(w, "try serializer.decrease_container_depth()")
+}
+
+fn push_deserializer(w: &mut dyn IndentWrite) -> io::Result<()> {
+    writeln!(w, "try deserializer.increase_container_depth()")
+}
+
+fn pop_deserializer(w: &mut dyn IndentWrite) -> io::Result<()> {
+    writeln!(w, "try deserializer.decrease_container_depth()")
+}
