--- crates/facet_generate/src/generation/swift/emitter/features/ListOfT.swift
@@ -1,22 +1,0 @@
-func serializeArray<T, S: Serializer>(
-    value: [T],
-    serializer: S,
-    serializeElement: (T, S) throws -> Void
-) throws {
-    try serializer.serialize_len(value: value.count)
-    for item in value {
-        try serializeElement(item, serializer)
-    }
-}
-
-func deserializeArray<T, D: Deserializer>(
-    deserializer: D,
-    deserializeElement: (D) throws -> T
-) throws -> [T] {
-    let length = try deserializer.deserialize_len()
-    var obj: [T] = []
-    for _ in 0..<length {
-        obj.append(try deserializeElement(deserializer))
-    }
-    return obj
-}
