--- crates/facet_generate/src/generation/swift/emitter/features/TupleArray.swift
@@ -1,21 +1,0 @@
-func serializeTupleArray<T, S: Serializer>(
-    value: [T],
-    serializer: S,
-    serializeElement: (T, S) throws -> Void
-) throws {
-    for item in value {
-        try serializeElement(item, serializer)
-    }
-}
-
-func deserializeTupleArray<T, D: Deserializer>(
-    deserializer: D,
-    size: Int,
-    deserializeElement: (D) throws -> T
-) throws -> [T] {
-    var obj: [T] = []
-    for _ in 0..<size {
-        obj.append(try deserializeElement(deserializer))
-    }
-    return obj
-}
